Updated JwtBearerExtentsions
